refactored home/welcome page behaviour
This makes the navbar solely dependent on the login state, so it's easier to
access the frontpage with upcoming events while you are logged in (I don't want
to punish logged-in users).
This also introduces a new pseudo theme, "horaro", which is what we use for
all Horaro pages from now on (both backend and frontend, but not event/schedule
frontends, which keep their theme-... classes).